Update postal response help question with link to upload question
I didn't realise we already addressed the large file upload question
so it makes more sense to link to that.

It's also probably not a good idea to put the P.O. Box option in
people's minds since that should be an absolutely last resort. The one
time it's been done so far it wasn't necessary anyway.