Remove outgoing mailer footer customisations
Our two customisations were asking FOI officers to not use PDF and
reminding them that requests from RTK are valid and even the
Information Commissioner has said so.

I'm suggesting we remove them since they're not having an impact.

The PDF one obviously isn't since we're still getting plenty of PDFs
and the valid request one has still seen instances of departments
playing silly buggers.

One technique that has been more effective is writing these statements
in the body of the message or in follow up messages. So let's continue
to do that where there's a problem but not codify it in the system.