slightly reduce the size of notices on wide screens
the 1.4em font-size on the notice makes the other text on
the page appear uncomfortably small and I think breaks the
heading hierarchy in an unhelpful way. I understand that the
notice is meant to stand out, but this is achieved by the box
style and colour, the extra large size is more noise than help.

This reduces the size to a more comfortable 1.2em, slightly
smaller than the h2 which I think makes it flow better.