Change tags used to determine public body jurisdiction
With our old tag selection there was no way to show all authorities
in a particular jurisdiction.

This way has the advantage of just a single tag being used to
determine jurisdiction. On the downside we now need another tag for
the category. However Federal authorities already need multiple tags
so it's obviously an OK trade off.