Explicitly specify path
This fixes a bug with Alaveteli's `script/load-mail-server-logs`. It
incorrectly sets the path it runs in to be the scripts directory instead
of the root. This means this bit of code throws an exception complaining
it can't find the locale directory.

This works around that problem by explicitly specifying the path, which is
a good idea anyway.