Make the padding of error lists more even
The current style relies on default browser <ul> styles.
This makes it quite uneven and not related to the rest of the box
padding—it sticks out too much on the left.

This commit makes the side margins of the box even,
and reduces the list padding so it sits evenly.