Make the 'quote' graphic lighter
The 'quote' graphic that sits with extracts from requests in lists,
and with the 'user description' on user profiles,
is very bold and visually dominates the page too much.

While it's a nice touch that brings interest and helps communicate
that these sections are a human voice, it's too heavy.

On user lists for example, the title of the request, status and parties
involved mostly provide more usual information to people browsing them
than the correspondence extract. But the heavy quote mark takes the focus.

This commit makes the quote graphic much less bold, reducing it's
contrast so it stands out less.

This means we keep the communication that this bit is the human voice,
and it makes the page a little more dynamic, but it fits better into the
page structure and allows the other important information to come
through.

Relates to #185