Remove link to volunteer for the OpenAustralia Foundation
This list of ways to help is extensive already and doing any of these
things would be a great help to the OpenAustralia Foundation. I think
this just muddies the water—better to keep people focused on this stuff
than link off to other OAF stuff that they may or may not be interested in.