Remove application fees table
Rather than constantly keep this up to date as fees routinely increase
I've decided it's best we remove it. Not only does removing it mean we
don't have to keep it up to date it also takes the focus away from the
specifics and that there are fees. In a future commit I'd like to add
a mention of informal requests.

Closes #634.