Update Contact to remove individual email address
We get a lot of people who send an email to Kat, Matthew and I individually instead of sending the email to the contact email address. Now that we use Front, this creates duplicates of the email which break the change.

This commit removes the emails/paragraphs in relation to names.