Merge in changes from main application
Note that there is a new official feature for asking questions about 
privacy. We're not using that for the time being because it looks like 
it works a little differently. However, it would be good to check in 
future versions if we could switch over to that instead to minimise our 
changes in the theme