Update factsheet link and description
I've updated the fact sheet link. However the fact sheets used to be for the public and now they are designed for Freedom of Information practitioners to provide agency staff with a broad overview of the Freedom of Information Act 1982 (FOI Act) and the steps required to be taken when an FOI request is received. I think this is the best information a user can read, because they know what the OAIC is sharing with the agency on how to deal with requests.